+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   - Fix events not being processed in a timely manner
   - Mark plugin as non-reloadable for https://github.com/elastic/logstash/pull/6499
   - Add ECS compatibility support v1 and v8
+  - Relax logstash-core-plugin-api runtime dependency to 1.60 - 2.99
   - Remove unneeded runtime dependency on 'stud'
 
 ## 0.9.1